Move xhprof/xhgui tests to own test file
